Add an extension that exposes async errors
XCTest doesn't handle errors well in async tests. When relying simply on `async throws`, Xcode displays the error on the method signature of the failing test, and only reports that “the operation couldn't be completed.”

Using this extension exposes the real message of the error. It would prefer if it also indicated the actual line that threw, but that has not been solved yet. Let's hope that it is possible to put this functionality in XCTest, and that it is updated soon.